added merging between self parsed and open food db

+++ backend/src/scraper/utils.rs
@@ -1,6 +1,6 @@
 use sha2::{Digest, Sha256};
 
-pub fn generate_product_id(source: &str, name: &str, brand: &str, serving_size: &i64) -> String {
+pub fn generate_product_id(source: &str, name: &str, brand: &str, serving_size: &i32) -> String {
     let input = format!(
         "{}:{}:{}:{}",
         source.to_lowercase(),
